pyproject.toml
[tool.poetry]
name = "nea-geofly-mapper"
version = "0.1.0-alpha"
description = "Download and stitch map tiles to a single photo from https://nea.geofly.eu/"
authors = ["fabyr", "Wonno"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
fastkml = "^1.4.0"
lxml = "^6.0.2"
requests = "^2.33.0"
rich = "^14.3.3"
pillow = "^12.1.1"
pyproj = "^3.7.2"
pygeoif = "^1.6.0"

[tool.poetry.group.test.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.1.0"
pylint = "^4.0.5"

[tool.poetry.group.optional.dependencies]
pyinstaller = "^6.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pylint (>=4.0.5,<5.0.0)"
]
